Move cmp_version stuff to _vendor
